refactor: Inject Module 디렉토리 변경
