+++ app/src/main/java/camp/nextstep/edu/github/MainViewModel.kt
@@ -12,7 +12,7 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor(
     private val gitHubRepository: GitHubRepository,
 ) : ViewModel() {
-    
+
     private val _githubRepositoryInfoLiveData = MutableLiveData<List<GitHubRepositoryInfo>>()
     val githubRepositoryInfoLiveData: LiveData<List<GitHubRepositoryInfo>> get() = _githubRepositoryInfoLiveData
