+++ data/src/main/java/camp/nextstep/edu/github/di/RetrofitServiceModule.kt
@@ -1,5 +1,7 @@
-package camp.nextstep.edu.github
+package camp.nextstep.edu.github.di
 
+import camp.nextstep.edu.github.GithubService
+import camp.nextstep.edu.github.RetrofitManager
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
